ready replacement, links work in speech & thought
