+++ public/style.css
@@ -15,7 +15,7 @@ a:visited{
 }
 span{
 font-size: inherit;
-margin:0.5rem;
+margin-block:0.5rem;
 height:inherit;
 }
